+++ server/services/postgresService.ts
@@ -1,5 +1,5 @@
 import { Task, TaskModel } from "../db/sequelizeModel";
-import { GenezioDeploy } from "@genezio/types";
+import { GenezioAuth, GenezioDeploy, GnzContext } from "@genezio/types";
 import { DataTypes, ModelStatic, Sequelize } from "sequelize";
 import { postgresURL } from "../config/envHandler";
 import pg from "pg";
@@ -7,7 +7,6 @@ import {
   CreateTaskRequestPostgres,
   CreateTaskResponse,
   GetTasksResponse,
-  UpdateTaskRequest,
   UpdateTaskRequestPostgres,
   UpdateTaskResponsePostgres,
 } from "../dtos/task";
@@ -56,11 +55,17 @@ export class PostgresService {
     return maxId + 1;
   }
 
+  @GenezioAuth()
   async createTask(
+    context: GnzContext,
     task: CreateTaskRequestPostgres
   ): Promise<CreateTaskResponse> {
     // Implementation for creating a task
+    const ownerId = context.user?.userId;
+    if (!ownerId) throw new Error("User not found in the context.");
     task.taskId = await this.#generateUniqueId();
+    task.date = new Date();
+    task.ownerId = ownerId;
     let createdTask: Task;
     try {
       createdTask = await this.model.create(task);
@@ -72,11 +77,16 @@ export class PostgresService {
     };
   }
 
-  async readTasks(): Promise<GetTasksResponse> {
+  @GenezioAuth()
+  async readTasks(context: GnzContext): Promise<GetTasksResponse> {
     // Implementation for reading tasks
+    const ownerId = context.user?.userId;
+    if (!ownerId) throw new Error("User not found in the context.");
     let tasks: Task[];
     try {
-      tasks = await this.model.findAll();
+      tasks = await this.model.findAll({
+        where: { ownerId: ownerId },
+      });
     } catch (error: any) {
       throw error;
     }
@@ -85,15 +95,20 @@ export class PostgresService {
     };
   }
 
+  @GenezioAuth()
   async updateTask(
+    context: GnzContext,
     updatedTask: UpdateTaskRequestPostgres
   ): Promise<UpdateTaskResponsePostgres> {
     // Implementation for updating a task
+    const ownerId = context.user?.userId;
+    if (!ownerId) throw new Error("User not found in the context.");
+
     updatedTask.date = new Date();
     let updatedTaskResponse;
     try {
       updatedTaskResponse = await this.model.update(updatedTask, {
-        where: { taskId: updatedTask.id },
+        where: { taskId: updatedTask.id, ownerId: ownerId },
       });
     } catch (error: any) {
       throw error;
@@ -103,10 +118,14 @@ export class PostgresService {
     };
   }
 
-  async deleteTask(taskId: number): Promise<void> {
+  @GenezioAuth()
+  async deleteTask(context: GnzContext, taskId: number): Promise<void> {
     // Implementation for deleting a task
+    const ownerId = context.user?.userId;
+    if (!ownerId) throw new Error("User not found in the context.");
+
     try {
-      await this.model.destroy({ where: { taskId: taskId } });
+      await this.model.destroy({ where: { taskId: taskId, ownerId: ownerId } });
     } catch (error: any) {
       throw error;
     }
